Refatorando servicos e criando crud

+++ configuracao/src/main/java/org/ecad/configuracao/repository/AmbienteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AmbienteRepository extends CrudRepository<Ambiente, Integer> {
+public interface AmbienteRepository extends CrudRepository<Ambiente, Long> {
 	
 }
