+++ configuracao/src/main/java/org/ecad/configuracao/repository/PropriedadeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropriedadeRepository extends CrudRepository<Propriedade, Integer> {
+public interface PropriedadeRepository extends CrudRepository<Propriedade, Long> {
 	
 }
