+++ configuracao/src/main/java/org/ecad/configuracao/repository/SistemaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SistemaRepository extends CrudRepository<Sistema, Integer> {
+public interface SistemaRepository extends CrudRepository<Sistema, Long> {
 	
 }
